// waspc/packages/studio/client/src/appSpec.ts
export type AppSpec = {
  decls: Decl[];
  waspProjectDir: string;
  isBuild: boolean;
  devEnvVarsServer: string[];
  devEnvVarsClient: string[];
  devDatabaseUrl: string | null;
  userDockerfileContents: string | null;
  customViteConfigPath: string | null;
  srcTsConfigPath: string;
  srcTsConfig: unknown;
  packageJson: unknown;
  tailwindConfigFilesRelocators: unknown;
  prismaSchema: unknown;
  migrationsDir: string | null;
  externalCodeFiles: unknown;
  externalPublicFiles: unknown;
};

export type Decl = {
  [Type in keyof DeclTypeToValue]: {
    declType: Type;
    declName: string;
    declValue: DeclTypeToValue[Type];
  };
}[keyof DeclTypeToValue];

export type DeclTypeToValue = {
  App: App;
  Page: Page;
  Route: Route;
  Query: Query;
  Action: Action;
  Job: Job;
  Api: Api;
  ApiNamespace: ApiNamespace;
  Crud: Crud;
  Entity: Entity;
};

export type GetDeclForType<T extends Decl["declType"]> = Extract<
  Decl,
  { declType: T }
>;